inject env only if it exist

+++ postbuild.sh
@@ -13,6 +13,7 @@ cp -r public ./.amplify-hosting/static
 
 cp deploy-manifest.json ./.amplify-hosting/deploy-manifest.json
 cp package.json ./.amplify-hosting/compute/default/package.json
-cp .env ./.amplify-hosting/compute/default/.env
+if [ -f .env ]; then
+    cp .env ./.amplify-hosting/compute/default/.env
+fi
 echo "NODE_ENV=production" >> ./.amplify-hosting/compute/default/.env
-# find . -maxdepth 1 -type f ! -name 'index.html' -exec cp {} ./.amplify-hosting/compute/default/ \;
